backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
asyncpg>=0.30.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP client (external API calls)
httpx>=0.27.0

# Config
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
